Get FujiNet version first before anything else

+++ localcmd/include/json.h
@@ -5,6 +5,6 @@
 
 extern uint8_t json_open(const char *path);
 extern void json_close();
-extern size_t json_query(const char *query, void *buffer);
+extern int json_query(const char *query, void *buffer);
 
 #endif /* JSON_H */
